Introduce a container to hold injection strategies
This also splits up all the default strategies (args, hash, and kwargs) into their own strategy classes.
