+++ lib/dry/auto_inject/strategies/constructor.rb
@@ -1,0 +1,67 @@
+require 'dry/auto_inject/dependency_map'
+
+module Dry
+  module AutoInject
+    class Strategies
+      class Constructor < Module
+        ClassMethods = Class.new(Module)
+        InstanceMethods = Class.new(Module)
+
+        attr_reader :container
+        attr_reader :dependency_map
+        attr_reader :instance_mod
+
+        def initialize(container, *dependency_names)
+          @container = container
+          @dependency_map = DependencyMap.new(*dependency_names)
+          @instance_mod = InstanceMethods.new
+        end
+
+        # @api private
+        def included(klass)
+          define_container(klass)
+          define_readers
+
+          define_new(klass)
+          define_initialize(klass)
+
+          klass.send(:include, instance_mod)
+
+          super
+        end
+
+        private
+
+        # @api private
+        def define_container(klass)
+          klass.instance_variable_set('@container', container)
+
+          klass.class_eval do
+            def self.container
+              if superclass.respond_to?(:container)
+                superclass.container
+              else
+                @container
+              end
+            end
+          end
+        end
+
+        def define_readers
+          instance_mod.class_eval <<-RUBY, __FILE__, __LINE__ + 1
+            attr_reader #{dependency_map.names.map { |name| ":#{name}" }.join(', ')}
+          RUBY
+          self
+        end
+
+        def define_new(klass)
+          raise NotImplementedError, "must be implemented by a subclass"
+        end
+
+        def define_initialize(klass)
+          raise NotImplementedError, "must be implemented by a subclass"
+        end
+      end
+    end
+  end
+end
